requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy>=1.23,<2.0
pandas>=1.5
pyarrow>=12.0
scikit-learn>=1.2

torch==2.2.2
faiss-cpu>=1.7.4

fastapi>=0.100,<1.0
uvicorn[standard]>=0.22,<1.0
pydantic>=2.0,<3.0

requests>=2.31
python-dotenv>=1.0
tqdm>=4.65
joblib>=1.3
typing-extensions>=4.7
